// wmf的代码/project/types.ts
export enum UserRole {
  Teacher = 'Teacher',             // 教师
  CollegeAdmin = 'CollegeAdmin',   // 二级学院管理员
  AssetAdmin = 'AssetAdmin',       // 资产处管理员
  Guest = 'Guest'                  // 游客
}